requirements.txt
google-genai
python-dotenv
requests
sentence-transformers
rank-bm25
numpy
